+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/bot/RequestException.kt
@@ -1,13 +1,14 @@
 package com.github.insanusmokrassar.TelegramBotAPI.bot
 
+import com.github.insanusmokrassar.TelegramBotAPI.bot.exceptions.RequestException
 import com.github.insanusmokrassar.TelegramBotAPI.types.Response
 import java.io.IOException
 
-class RequestException(
-    val response: Response<*>,
-    message: String? = null,
-    cause: Throwable? = null
-) : IOException(
-    message,
-    cause
+@Deprecated(
+    "Replaced to another package",
+    ReplaceWith(
+        "RequestException",
+        "com.github.insanusmokrassar.TelegramBotAPI.bot.exceptions.RequestException"
+    )
 )
+typealias RequestException = RequestException