+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Update some types and requests according to abstractions replacing
 * Add all `InlineQueryResult`, `InputMessageContent` and other inline mode types
 * Fixes in edition of inline messages and their result types
+* Replace basic exception and add `ReplyMessageNotFound` exception
 
 ## 0.9.0